requirements.txt
# Requerimientos para el backend Django
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9.6  # Si se usa PostgreSQL

# Autenticación opcional
django-allauth>=0.54

# Seguridad y control de permisos
django-guardian>=2.4.0

# Almacenamiento (para futuros informes PDF o archivos)
django-storages>=1.14

# Para manejar variables de entorno (si se usa .env)
python-decouple>=3.8

# FRONTEND (Eliana)
# Instalar con npm:
# npm install vue@3
# npm install -D tailwindcss postcss autoprefixer
# npx tailwindcss init -p
